# fix(engine): Partially fix non-resolvable-interface-object