--- a/crates/engine/Cargo.toml
+++ b/crates/engine/Cargo.toml
@@ -26,6 +26,7 @@ bytes.workspace = true
 url.workspace = true
 crossbeam-queue = "0.3.11"
 futures-util.workspace = true
+fixedbitset.workspace = true
 futures.workspace = true
 grafbase-telemetry.workspace = true
 hex.workspace = true