--- a/crates/engine/src/response/write/deserialize/object/concrete.rs
+++ b/crates/engine/src/response/write/deserialize/object/concrete.rs
@@ -31,7 +31,7 @@ impl<'ctx, 'seed> ConcreteShapeSeed<'ctx, 'seed> {
                 has_error: shape.has_errors(),
                 object_identifier: shape.identifier,
                 field_shape_ids: shape.field_shape_ids,
-                typename_response_edges: &shape.as_ref().typename_response_edges,
+                typename_response_edges: &shape.as_ref().typename_response_keys,
             },
         }
     }
@@ -50,7 +50,7 @@ impl<'ctx, 'seed> ConcreteShapeSeed<'ctx, 'seed> {
                 has_error: shape.has_errors(),
                 object_identifier: ObjectIdentifier::Known(object_id),
                 field_shape_ids: shape.field_shape_ids,
-                typename_response_edges: &shape.as_ref().typename_response_edges,
+                typename_response_edges: &shape.as_ref().typename_response_keys,
             },
         }
     }
@@ -164,7 +164,7 @@ impl<'de, 'ctx, 'seed> Visitor<'de> for ConcreteShapeFieldsSeed<'ctx, 'seed> {
 
         if !self.typename_response_edges.is_empty() {
             let Some(object_id) = maybe_object_id else {
-                return Err(serde::de::Error::custom("Could not determine the "));
+                return Err(serde::de::Error::custom("Could not determine the object type"));
             };
             let name_id = schema[object_id].name_id;
             for edge in self.typename_response_edges {