--- a/crates/integration-tests/tests/federation/basic/mod.rs
+++ b/crates/integration-tests/tests/federation/basic/mod.rs
@@ -3,7 +3,6 @@
 //! This file shouldn't have much federation specific stuff in it, mostly just checking
 //! that our engine supports all the things a normal GraphQL server should.
 
-// mod caching;
 mod empty_config;
 mod enums;
 mod errors;
@@ -13,6 +12,7 @@ mod mutation;
 mod operation_limits;
 mod operations;
 mod scalars;
+mod shapes;
 mod skip_include;
 mod streaming;
 mod typename;