--- a/crates/integration-tests/tests/federation/basic/shapes/mod.rs
+++ b/crates/integration-tests/tests/federation/basic/shapes/mod.rs
@@ -1,0 +1,1 @@
+mod interface;