… Cargo.lock
… Cargo.toml
… crates/engine/src/operation/solve/solver/shapes/mod.rs
… crates/engine/src/operation/solve/solver/shapes/partition.rs
… crates/integration-tests/benches/federation/snapshots/bench__federation__introspection__prepare-2.snap
… crates/integration-tests/benches/federation/snapshots/bench__federation__introspection__prepare.snap
… crates/integration-tests/tests/federation/basic/caching.rs
… crates/integration-tests/tests/federation/basic/shapes/interface/one_interface.rs
… crates/integration-tests/tests/federation/basic/shapes/interface/one_union.rs
… crates/integration-tests/tests/federation/basic/shapes/interface/only_objects.rs
… crates/integration-tests/tests/federation/basic/shapes/interface/two_interfaces.rs