So the primary goal of this PR was to solve the `non-resolvable-interface-object` test, which it improves. The problem was that we query:
```graphql
query {
  node {
     id
  }
}
```

with a schema like:

```graphql
type Query {
  node: Node!
}

interface Node {
  id: ID!
}
```

Not that `Node` has no implementors but the subgraph still returns data. However, with the current design when I needed to have at least one implementor.

Let's take a step back, to deserialize data I use extensively `DeserializeSeed` to avoid any extra allocations like `serde_json::Value` would. But this means that I need to know what kind of data I expect during deserialization. To do this we construct all the possible field and object shapes we may encounter. This is potentially an extremely costly step so I had already made some effort in an efficient partitioning logic to partition the possible output types into distinct partitions that each have the same set of fields and thus the same concrete shape. If we have:
```graphql
interface Node {
  id: ID!
}

interface Other {
  other: String!
}

type A implements Node & Other { ... }
type B implements Node { .. }

type Query {
  node: Node!
}
```

then `{ node { id ... on Other { other } } }` has two distinct partitions: `[[A], [B]]` as they each have different fields. but if we only query `{ node { id } }` then we have a single partition `[A, B]`.

The current logic didn't keep track of the associated fields with each partition though. So I used any object id in the partition and iterated over all fields filtering out the ones that don't apply. And here we come back to the `non-resolvable-interface-object` test, as there wasn't any implementors, we didn't have any object and couldn't create the field shape!

So to fix that and improve the performance, as field filtering was slow, I'm now keeping track of fields in a bitset. This significantly improved performance and for fun I removed the other significant performance problem which was keeping track of which field shape was associated with which field. Now I'm avoid most allocations for this and using something similar to a counting sort. Both of those reduce by a 10x factor the time it takes for operation preparation in my worst-case scenario (440ms -> ~40ms) and 3x for the simplest bench I have for this (1ms -> 300us). So small extra win.

While the partitioning logic is tested with unit-tests I didn't have any decent tests at the GraphQL layer including how shapes are used and not only generated. So wrote a bunch of those. And hit a corner case with fragments...
For a schema like:
```graphql
interface Node {
  id: ID!
}

interface Other {
  other: String!
}

type A implements Node & Other { ... }
type B implements Node { .. }
type C implements Other { .. }

type Query {
  node: Node!
}
```
It is valid to write:
```graphql
query {
  node {
    ...Frag
  }
}

fragment Frag on Other {
   ... on C { other }
}
```

But this ends up as: `query { node { ... on C { other } } }` in the engine as we flatten all type conditions. We weren't removing those fields and instead were requesting those in subgraph queries... So did a quick hack to solve it, filtering them out in the binding step. But will move it to the planning step and simplify the binding step to keep fragments later.